Add files via upload
Established a node capable of interfacing with PC so that a user may enter text commands. Based on the commands, subsequent transmission on the 2-wire RS-485 bus to other nodes will occur. The node will also extract information out of the asynchronous data stream and to control three or more devices.